Do not leave live nodes disconnected from entry point after removing deleted nodes
This is an edge case we hit in production recently.

It might be the case that we would leave live nodes disconnected from
the entry point for certain graphs, after removing deleted nodes.

As a reproducer, consider the following valid graph with max degree
of 2.

<img width="605" alt="image" src="https://github.com/user-attachments/assets/fab2f406-c60f-4885-8dbb-49b9c76e2bf1">

When all the nodes reachable from old entry point is deleted, we are
left with a live, but disconnected node. The current code sets the
entry point to -1 for this case, and if you were to call cleanup(instead of
remeoveDeletedNodes) here, you would get a totally different error
as we expect to have a valid entry node when the graph size is greater
than 0.